+++ resolwe_bio/utils/test.py
@@ -58,7 +58,8 @@ class BioProcessTestCase(ProcessTestCase):
                   "bowtie_index": "bt_index.tar.gz",
                   "bowtie2_index": "bt2_index.tar.gz",
                   "bwa_index": "bwa_index.tar.gz",
-                  "hisat2_index": "hisat2_index.tar.gz"}
+                  "hisat2_index": "hisat2_index.tar.gz",
+                  "subread_index": "subread_index.tar.gz"}
         return self.run_process('upload-genome', inputs)
 
     def prepare_reads(self, fn=['reads.fastq.gz']):