+++ docs/CHANGELOG.rst
@@ -33,6 +33,7 @@ Added
 * Add upload processor for common image file types (.jpg/.tiff/.png/.gif)
 * Add upload processor for tabular file formats (.tab/.tsv/.csv/.txt/.xls/.xlsx)
 * Add featureCounts process
+* Add Subread process
 
 
 ================